% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caption_db.R
\docType{data}
\name{caption_db}
\alias{caption_db}
\title{My caption for plots}
\format{
An object of class \code{labels} of length 1.
}
\usage{
caption_db
}
\description{
This saves me writing one line of code, everytime.
}
\keyword{datasets}
