% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_util.r
\name{convert.db.to.r}
\alias{convert.db.to.r}
\title{Convert data from a database table to R format}
\usage{
convert.db.to.r(vals, rclass = schema$rclass, schema = NULL,
  as.data.frame = is.data.frame(vals))
}
\arguments{
\item{vals}{the values loaded from the database table}

\item{rclass}{the r class of the table columns, is extracted from schema}

\item{schema}{a table schema that can be used to convert values}
}
\description{
Convert data from a database table to R format
}

