% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_util.r
\name{dbConnectSQLiteWithSchema}
\alias{dbConnectSQLiteWithSchema}
\title{Connect to an SQLite db and set
a schema}
\usage{
dbConnectSQLiteWithSchema(
  dbname,
  schema.file = paste0(tools::file_path_sans_ext(dbname), ".yaml"),
  schema = load.and.init.schemas(schema.file)
)
}
\description{
Connect to an SQLite db and set
a schema
}
