% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_util.r
\name{dbCreateSQLiteFromSchema}
\alias{dbCreateSQLiteFromSchema}
\title{Create a new SQLite database from a schema file}
\usage{
dbCreateSQLiteFromSchema(schema.file, schema.dir = dirname(schema.file),
  db.name = NULL, db.dir = schema.dir)
}
\arguments{
\item{schema.file}{the dbmisc schema file in yaml format}

\item{schema.dir}{the directory of the schema file (if schema.file does not contain a path)}

\item{db.name}{the name of the database file}

\item{db.dir}{the directory of the database file, by default the schema directory}
}
\description{
Create a new SQLite database from a schema file
}

