% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_util.r
\name{dbDelete}
\alias{dbDelete}
\title{Delete row(s) from table}
\usage{
dbDelete(conn, table, params, sql = NULL, run = TRUE)
}
\arguments{
\item{conn}{dbi database connection}

\item{table}{name of the table}

\item{params}{named list of values for key fields that identify the rows to be deleted}

\item{sql}{optional a parameterized sql string}

\item{run}{if FALSE only return parametrized SQL string}
}
\description{
Delete row(s) from table
}

