% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_util.r
\name{dbGetMemoise}
\alias{dbGetMemoise}
\title{Get results from a database like dbGet put buffer the results in memory}
\usage{
dbGetMemoise(db, table, params = NULL, schema = schemas[[table]],
  schemas = get.db.schemas(db), log.dir = NULL,
  refetch.if.changed = !is.null(log.dir), empty.as.null = FALSE)
}
\description{
If the function is called again with the same parameter check if the
something was changed in the database inbetween by looking at the time
stamp of the log file. If there were no changes restore the values from
memory. If there were changes load data again from database.
}
\details{
If refetch.if.changed = FALSE (default if no log.dir is provided), always
use the data from memory.
}

