% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_util.r
\name{dbInsert}
\alias{dbInsert}
\title{Insert row into table}
\usage{
dbInsert(conn, table, vals, sql = NULL, run = TRUE, mode = c("insert",
  "replace")[1])
}
\arguments{
\item{conn}{dbi database connection}

\item{table}{name of the table}

\item{vals}{named list of values to be inserted}

\item{sql}{optional a parameterized sql string}

\item{run}{if FALSE only return parametrized SQL string}

\item{mode}{"insert" or "replace", should have no effect so far}
}
\description{
Insert row into table
}

