% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_util.r
\name{dbTableCols}
\alias{dbTableCols}
\title{Get a data frame with column information for a database table}
\usage{
dbTableCols(db, table)
}
\arguments{
\item{db}{dbi database connection}

\item{table}{name of the table}
}
\description{
Get a data frame with column information for a database table
}
