% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_util.r
\name{dbUpdate}
\alias{dbUpdate}
\title{Update a row in a database table}
\usage{
dbUpdate(conn, table, vals, where = NULL, schema = NULL, sql = NULL,
  run = TRUE, rclass = schema$rclass, convert = !is.null(rclass),
  null.as.na = TRUE)
}
\arguments{
\item{conn}{dbi database connection}

\item{table}{name of the table}

\item{vals}{named list of values to be inserted}

\item{where}{named list that specifies the keys where to update}

\item{schema}{a schema as R list, can be used to automatically convert types}

\item{sql}{optional a parameterized sql string}

\item{run}{if FALSE only return parametrized SQL string}

\item{rclass}{the r class of the table columns, is extracted from schema}

\item{convert}{if rclass is given shall results automatically be converted to these classes?}

\item{null.as.na}{shall NULL values be converted to NA values?}
}
\description{
Update a row in a database table
}

