% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_util.r
\name{empty.df.from.schema}
\alias{empty.df.from.schema}
\title{Creates an example data frame from a database schema table
using provided column values and default values specified in schema}
\usage{
empty.df.from.schema(.schema, .nrows = 1, ..., .use.defaults = TRUE)
}
\description{
Creates an example data frame from a database schema table
using provided column values and default values specified in schema
}
