% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_xaringan.R
\name{forest_xaringan}
\alias{forest_xaringan}
\title{Forest Research Xaringan Theme}
\usage{
forest_xaringan()
}
\description{
This file contains a pre-designated theme which can be easily called as a function
}
