% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{forestgreen}
\alias{forestgreen}
\title{Forest Research Colour Palette}
\format{
An object of class \code{character} of length 1.
}
\usage{
forestgreen
}
\description{
This file contains the colours we will use for our ggplot2 and xaringan theme
}
\keyword{datasets}
