% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_util.r
\name{init.schema}
\alias{init.schema}
\title{Init a schema by parsing table definition and store info
in easy accessibale R format}
\usage{
init.schema(schema, name = NULL)
}
\arguments{
\item{schema}{the table schema as an R list}

\item{name}{of the table}
}
\description{
Create rclasses of each column and primary keys
}
