% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_util.r
\name{load.and.init.schemas}
\alias{load.and.init.schemas}
\title{Load and init database table schemas from yaml file}
\usage{
load.and.init.schemas(file = NULL, yaml = NULL)
}
\arguments{
\item{file}{file name}

\item{yaml}{yaml as text}
}
\description{
Load and init database table schemas from yaml file
}

