% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_util.r
\name{logDBcommand}
\alias{logDBcommand}
\title{log a command that changes a database}
\usage{
logDBcommand(
  type,
  sql = "",
  user = "NA",
  log.dir = NULL,
  table = NULL,
  do.log = TRUE
)
}
\description{
log a command that changes a database
}
