% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_util.r
\name{schema.r.classes}
\alias{schema.r.classes}
\title{Get a vector of R classes of the database columns described in a schema}
\usage{
schema.r.classes(schema)
}
\arguments{
\item{schema}{the schema}
}
\description{
Get a vector of R classes of the database columns described in a schema
}
