% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_util.r
\name{schema.template}
\alias{schema.template}
\title{Create an example schema from a list of R objects}
\usage{
schema.template(li, name = "mytable", toClipboard = TRUE)
}
\arguments{
\item{li}{The R list for which the schema shall be created}

\item{name}{optional a name of the table}

\item{toCliboard}{shall the created text be copied to the clipboard}
}
\description{
The output is shown per cat and copied to the clipboard.
It can be used as a template for the .yaml schema file
}
