% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_util.r
\name{set.db.schemas}
\alias{set.db.schemas}
\title{Set schemas as hidden attribute to a data base connection db}
\usage{
set.db.schemas(db, schemas = NULL, schema.file = NULL)
}
\description{
Set schemas as hidden attribute to a data base connection db
}

