% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_util.r
\name{sql.where.code}
\alias{sql.where.code}
\title{Create a parametrized or escaped SQL WHERE clause from
the provided parameters.}
\usage{
sql.where.code(
  db = NULL,
  params,
  where.in = FALSE,
  parametrized = !where.in,
  add.where = TRUE
)
}
\arguments{
\item{db}{a database connection needed for correct escaping via glue_sql}

\item{params}{a list of parameters assume that db fields have the same name and have to be equal to provided values.}

\item{where.in}{Set true TRUE a member of params can be vector and return all rows that match an element. By default FALSE to generate more compact code.}

\item{add.where}{If TRUE start with WHERE}

\item{paramertrized}{shall the generated code use SQL parameters.}
}
\description{
Create a parametrized or escaped SQL WHERE clause from
the provided parameters.
}
