% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_forest.R
\name{theme_forest}
\alias{theme_forest}
\title{Custom ggplot Theme in Forest Research Flavor.}
\usage{
theme_forest()
}
\description{
This is a minimal change to the minimal theme of ggplot2, to position text elements and add some colour.
}
