% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_util.r
\name{to.db.datetime}
\alias{to.db.datetime}
\title{Convert an R object to a datetime object that can be used
in a WHERE clause.}
\usage{
to.db.datetime(val, origin = "1970-01-01")
}
\description{
Convert an R object to a datetime object that can be used
in a WHERE clause.
}
