% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_util.r
\name{tp.db.date}
\alias{tp.db.date}
\title{Convert an R object to a date object that can be used
in a WHERE clause.}
\usage{
tp.db.date(val, origin = "1970-01-01")
}
\description{
Convert an R object to a date object that can be used
in a WHERE clause.
}
