% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_forest.R
\docType{data}
\name{caption_db}
\alias{caption_db}
\title{Custom ggplot Theme in Forest Research Flavor.}
\format{
An object of class \code{labels} of length 1.
}
\usage{
caption_db
}
\description{
This is a minimal change to the minimal theme of ggplot2, to position text elements and add some colour.
}
\keyword{datasets}
