% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_util.r
\name{dbCreateSchemaTables}
\alias{dbCreateSchemaTables}
\title{Create database tables and possible indices from a simple yaml schema}
\usage{
dbCreateSchemaTables(db, schemas = get.db.schemas(db), schema.yaml = NULL,
  schema.file = NULL, overwrite = FALSE, silent = FALSE)
}
\arguments{
\item{db}{dbi database connection}

\item{schemas}{schemas as R list}

\item{schema.yaml}{alternatively a schema as yaml text}

\item{schema.file}{alternatively a file name of a schema yaml file}

\item{overwrite}{shall existing tables be overwritten?}

\item{silent}{if TRUE don't show messages}
}
\description{
Create database tables and possible indices from a simple yaml schema
}

