% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_modify.R
\name{dbCreateSQLiteFromSchema}
\alias{dbCreateSQLiteFromSchema}
\title{Create or update a SQLite database from a schema file}
\usage{
dbCreateSQLiteFromSchema(schema.file, schema.dir = dirname(schema.file),
  db.name = NULL, db.dir = schema.dir, update = TRUE)
}
\arguments{
\item{schema.file}{the dbmisc schema file in yaml format}

\item{schema.dir}{the directory of the schema file (if schema.file does not contain a path)}

\item{db.name}{the name of the database file}

\item{db.dir}{the directory of the database file, by default the schema directory}

\item{update}{if TRUE (default) copy and update the existing data in the tables. If FALSE just generate empty tables}
}
\description{
Create or update a SQLite database from a schema file
}

