% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_util.r
\name{dbGet}
\alias{dbGet}
\title{Get rows from a table}
\usage{
dbGet(db, table = NULL, params = NULL, sql = NULL, run = TRUE,
  schema = NULL, rclass = schema$rclass, convert = !is.null(rclass),
  orderby = NULL, null.as.na = TRUE, origin = "1970-01-01")
}
\arguments{
\item{db}{dbi database connection}

\item{table}{name of the table}

\item{params}{named list of values for key fields that identify the rows to be deleted}

\item{sql}{optional a parameterized sql string
       if you want to insert into the sql string
       the value from a provided parameter mypar
       write :mypar in the SQL string at the corresponding position.
       Example:

       select * from mytable where name = :myname}

\item{run}{if FALSE only return parametrized SQL string}

\item{schema}{a table schema that can be used to convert values}

\item{rclass}{the r class of the table columns, is extracted from schema}

\item{convert}{if rclass is given shall results automatically be converted to these classes?}

\item{orderby}{names of columns the results shall be ordered by as character vector. Add "DESC" or "ASC" after column name to sort descending or ascending. Example: `orderby = c("pos DESC","hp ASC")`}
}
\description{
Get rows from a table
}

