% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_util.r
\name{dbGet}
\alias{dbGet}
\title{Get rows from a table}
\usage{
dbGet(.db, .table = NULL, ..., .par = NULL, .sql = NULL, .run = TRUE)
}
\arguments{
\item{.db}{dbi database connection}

\item{.table}{name of the table}

\item{...}{alternatively parameters can be provided directly as named arguments}

\item{.par}{named list of values for key fields that identify the rows to be deleted}

\item{.sql}{optional a parameterized sql string
       if you want to insert into the sql string
       the value from a provided parameter mypar
       write :mypar in the SQL string at the corresponding position.
       Example:

       select * from mytable where name = :myname}

\item{.run}{if FALSE only return parametrized SQL string}
}
\description{
Get rows from a table
}

